requirements.txt
Django>=5.2
psycopg2-binary
djangorestframework
djangorestframework-simplejwt
drf_yasg